Fix triattention_gpu_gather_k_rows stub signature parameter type mismatch

+++ src/llama-triattention-gpu-stub.cpp
@@ -94,6 +94,6 @@ bool triattention_gpu_gather_k_rows(
     size_t,
     const uint32_t *,
     uint32_t,
-    size_t) {
+    uint32_t) {
     return false;
 }
